requirements.txt
agate==1.12.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
annotated-types==0.7.0
attrs==24.3.0
babel==2.16.0
boto3==1.34.0
botocore==1.34.162
cached-property==2.0.1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
csv_detective==0.7.4
csvkit==2.0.1
cycler==0.12.1
dacite==1.8.1
dataflows-tabulator==1.54.3
DateTime==5.5
dbfread==2.0.7
defusedxml==0.7.1
et_xmlfile==2.0.0
faust-cchardet==2.1.19
fonttools==4.55.3
frformat==0.3.0
greenlet==3.1.1
htmlmin==0.1.12
idna==3.10
ijson==3.3.0
ImageHash==4.3.1
importlib_resources==6.5.2
iniconfig==2.0.0
isodate==0.7.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json_extract==1.1.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
leather==0.4.0
linear-tsv==1.1.0
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.4
missingno==0.5.2
multimethod==1.12
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandas-dq==1.29
parsedatetime==2.6
pathlib==1.0.1
patsy==1.0.1
phik==0.12.4
pillow==11.0.0
pluggy==1.5.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
pytest==8.3.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
python-stdnum==1.20
pytimeparse==1.1.8
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
responses==0.25.0
rfc3986==2.0.0
rpds-py==0.22.3
s3transfer==0.9.0
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
setuptools==75.6.0
shapely==2.0.6
six==1.17.0
SQLAlchemy==2.0.36
statsmodels==0.14.4
sweetviz==2.3.1
tableschema==1.21.0
tangled-up-in-unicode==0.2.0
text-unidecode==1.3
threadpoolctl==3.5.0
tqdm==4.67.1
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2024.2
unicodecsv==0.14.1
Unidecode==1.3.6
urllib3==2.3.0
visions==0.7.6
wordcloud==1.9.4
xlrd==2.0.1
ydata-profiling==4.12.1
zope.interface==7.2
